+++ ninja-xtask/src/lib.rs
@@ -18,145 +18,6 @@ pub mod exit_with_json;
 
 pub use exit_with_json::{Exit, WithJson};
 
-#[derive(Debug)]
-pub struct Cmd {
-    pub name: &'static str,
-    pub result: Result<Output, io::Error>,
-    pub flags: CheckFlags,
-}
-
-trait CmdExt {
-    fn into_cmd(self, name: &'static str, checkflags: Option<CheckFlags>) -> Cmd;
-}
-
-impl CmdExt for Result<Output, io::Error> {
-    fn into_cmd(self, name: &'static str, checkflags: Option<CheckFlags>) -> Cmd {
-        Cmd {
-            name,
-            result: self,
-            flags: checkflags.unwrap_or_default(),
-        }
-    }
-}
-
-impl From<Cmd> for Exit<WithJson<()>> {
-    fn from(cmd: Cmd) -> Self {
-        let flags = cmd.flags;
-        let task = cmd.name;
-        match cmd.result {
-            Ok(output) if flags.contains(CheckFlags::JSON) => {
-                let status = output.status;
-                let payload = String::from_utf8_lossy(&output.stdout)
-                    .lines()
-                    .filter(|line| line.starts_with("{"))
-                    .map(serde_json::from_str::<Value>)
-                    .map(|json| json.unwrap_or_else(|err| json!({"unparsable": &err.to_string()})))
-                    .collect::<Value>();
-                let json = json!({
-                    "task": task,
-                    "status": &status.to_string(),
-                    "payload": payload,
-                });
-                if status.success() {
-                    Exit::Ok(WithJson {
-                        value: (),
-                        json: Some(json),
-                    })
-                } else {
-                    Exit::Error(WithJson {
-                        value: String::new(),
-                        json: Some(json),
-                    })
-                }
-            }
-            Ok(output) if !output.status.success() => {
-                let stdout = String::from_utf8_lossy(&output.stdout);
-                let stderr = String::from_utf8_lossy(&output.stderr);
-                Self::Error(WithJson {
-                    value: format!(
-                        "====== {task} exited with {status} ======\n-- stdout: --\n{stdout}\n\n-- stderr: --\n{stderr}",
-                        status = output.status
-                    ),
-                    json: None,
-                })
-            }
-            Ok(_) => {
-                println!("{task}: OK");
-                Self::Ok(WithJson {
-                    value: (),
-                    json: None,
-                })
-            }
-            Err(err_spawning) if flags.contains(CheckFlags::JSON) => {
-                let json = json!({
-                    "task": task,
-                    "status": "failed to spawn",
-                    "error": &err_spawning.to_string(),
-                });
-                Exit::IO(WithJson {
-                    value: String::new(),
-                    json: Some(json),
-                })
-            }
-            Err(err_spawning) => {
-                let msg = format!("{task} failed: {err_spawning}");
-                Self::IO(WithJson {
-                    value: msg,
-                    json: None,
-                })
-            }
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct Spawned {
-    pub name: &'static str,
-    pub child: Result<Child, io::Error>,
-    pub flags: CheckFlags,
-}
-
-impl Spawned {
-    pub fn wait(self) -> Cmd {
-        match self.child {
-            Ok(child) => child
-                .wait_with_output()
-                .into_cmd(self.name, Some(self.flags)),
-            Err(e) => Cmd {
-                name: self.name,
-                result: Err(e),
-                flags: self.flags,
-            },
-        }
-    }
-}
-
-trait SpawnedExt {
-    fn into_spawned(self, name: &'static str, flags: Option<CheckFlags>) -> Spawned;
-}
-
-impl SpawnedExt for Result<Child, io::Error> {
-    fn into_spawned(self, name: &'static str, flags: Option<CheckFlags>) -> Spawned {
-        Spawned {
-            name,
-            child: self,
-            flags: flags.unwrap_or_default(),
-        }
-    }
-}
-
-impl FromIterator<Spawned> for Exit<WithJson<()>> {
-    fn from_iter<I: IntoIterator<Item = Spawned>>(spawns: I) -> Self {
-        spawns.into_iter().map(Exit::from).collect()
-    }
-}
-
-impl From<Spawned> for Exit<WithJson<()>> {
-    fn from(spawn: Spawned) -> Self {
-        spawn.wait().into()
-    }
-}
-
 #[derive(Parser)]
 #[command(name = "cargo")]
 #[command(bin_name = "cargo")]
@@ -214,53 +75,4 @@ impl From<&NinjaCommand> for CheckFlags {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use std::{assert_matches, process::Command};
 
-    use super::*;
-
-    #[test]
-    fn exit_from_404() {
-        let splat: Cmd = Command::new("splat").output().into_cmd("splat", None);
-        assert_eq!(splat.name, "splat");
-        assert!(
-            matches!(splat.result, Result::Err(ref e) if matches!(e.kind(), io::ErrorKind::NotFound))
-        );
-        let exit: Exit<WithJson<()>> = Exit::from(splat);
-        let Exit::IO(WithJson {
-            value: msg,
-            json: _,
-        }) = exit
-        else {
-            panic!("not an IO2")
-        };
-        eprintln!("{}", msg);
-        assert!(msg.starts_with("splat failed: "));
-    }
-
-    #[test]
-    fn collect_exit() {
-        let exits = [
-            Exit::Ok(WithJson {
-                value: (),
-                json: None,
-            }),
-            Exit::IO(WithJson {
-                value: "one\n".to_string(),
-                json: None,
-            }),
-            Exit::Error(WithJson {
-                value: "two\n".to_string(),
-                json: None,
-            }),
-            Exit::Error(WithJson {
-                value: "three\n".to_string(),
-                json: None,
-            }),
-        ];
-        let exit: Exit<WithJson<()>> = exits.into_iter().collect();
-        let expected = "one\ntwo\nthree\n";
-        assert_matches!(exit, Exit::IO(s) if s.value == expected);
-    }
-}
